+++ src/apps/seelen_rofi/index.tsx
@@ -9,7 +9,7 @@ import { getRootContainer } from '../shared';
 import { wrapConsole } from '../shared/ConsoleWrapper';
 import { App } from './App';
 import { registerDocumentEvents } from './events';
-import i18n from './i18n';
+import i18n, { loadTranslations } from './i18n';
 
 import './styles/reset.css';
 import './styles/colors.css';
@@ -17,6 +17,7 @@ import './styles/colors.css';
 async function Main() {
   wrapConsole();
   await declareDocumentAsLayeredHitbox();
+  await loadTranslations();
   await initStore();
   registerDocumentEvents();